% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssMousetrack-package.R
\docType{package}
\name{ssMousetrack-package}
\alias{ssMousetrack-package}
\alias{ssMousetrack}
\title{Bayesian State-Space Modeling of Mouse-Tracking Experiments Via Stan}
\description{
The \pkg{ssMousetrack} package allows analysing mouse-tracking experiments via Bayesian state-space modeling. 
The package estimates the model using Markov Chain Monte Carlo, variational approximations to the posterior distribution, or optimization, 
as implemented in the \pkg{rstan} package. The user can use the customary R modeling syntax to define equations of the model and Stan syntax to
specify priors over the model parameters. 

The sections below provide an overview of the state-space model implemented by the \pkg{ssMousetrack} package.
}
\references{
Calcagnì, A., Coco, M., Pastore, M., & Duran N. (2019). State space modeling for incidental memory in naturalistic scenes. \emph{Manuscript in preparation}

Calcagnì, A., Lombardi, L., & D'Alessandro, M. (2018). A state space approach to dynamic modeling of mouse-tracking data. \emph{Under review}

Calcagnì, A., Lombardi, L., & D'Alessandro (2018, August). Probabilistic modeling of mouse-tracking data: A statespace approach. Paper presented at the \emph{2018 European Mathematical Psychology Group Meeting} (EMPG 2018), Genova, Italy

Calcagnì, A., Lombardi, L., D'Alessandro, M., & Sulpizio S. (2018, March). A subject oriented state-space approach to model mouse-tracking data. Paper presented at the \emph{60th Conference of Experimental Psychologists} (TeaP 2018), Marburg, Germany

Freeman, J. B. (2018). Doing psychological science by hand. \emph{Current Directions in Psychological Science}, In press, 1-9

Särkkä, S. (2013). Bayesian Filtering and Smoothing. \emph{Cambridge University Press}

Durbin, J., & Koopman, S. J. (2012). Time series analysis by state space methods (Vol. 38). \emph{Oxford University Press}

Andrieu, C., Doucet, A., & Holenstein, R. (2010). Particle markov chain monte carlo methods. \emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}, 72(3), 269s-342

Gelman, A., Carlin, J. B., Stern, H. S., & Rubin, D. B. (2004). Bayesian Data Analysis (Second edition). Chapman & Hall/CRC.
}
\seealso{
\url{http://mc-stan.org/} for more information on the Stan C++ language used by \pkg{ssMousetrack} package

Jokkala, J. (2016). Github repository: \emph{kalman-stan-randomwalk}, \url{https://github.com/juhokokkala/kalman-stan-randomwalk}
}
